backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
groq==0.11.0
langgraph==0.2.34
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5
langchain-text-splitters==0.3.2
faiss-cpu==1.13.2
sentence-transformers==3.1.1
httpx==0.27.2
beautifulsoup4==4.12.3
pydantic==2.9.2
